// infrastructure/microservices/file-microservice/src/Domain/enums/AllowedFileExtensions.ts
export enum AllowedFileExtensions {
  // Image extensions for Animation Worker
  JPG = ".jpg",
  JPEG = ".jpeg", 
  PNG = ".png",
  GIF = ".gif",
  BMP = ".bmp",
  WEBP = ".webp",
  SVG = ".svg",
  
  // Video extensions for Animation Worker
  MP4 = ".mp4",
  AVI = ".avi",
  MOV = ".mov",
  WMV = ".wmv",
  FLV = ".flv",
  WEBM = ".webm",
  MKV = ".mkv",
  
  // Audio extensions for Audio & Music Stagist
  MP3 = ".mp3",
  WAV = ".wav",
  FLAC = ".flac",
  AAC = ".aac",
  OGG = ".ogg",
  M4A = ".m4a",
  WMA = ".wma"
}